Refactor flags save handler
