+++ app/dashboard/site/save/index.js
@@ -2,6 +2,7 @@ module.exports = {
   analytics: require("./analytics"),
   avatar: require("./avatar"),
   finish: require("./finish"),
+  flags: require("./flags"),
   format: require("./format"),
   injectTitle: require("./injectTitle"),
   removeTmpFiles: require("./removeTmpFiles"),
