+++ app/views/dashboard/site/settings/flags.html
@@ -3,6 +3,8 @@
 <h1>Flags</h1>
 <p>Toggle experimental features for this site.</p>
 
+<input type="hidden" name="redirect" value="{{{base}}}/settings/flags" />
+
 {{#flags}}
 <div style="margin: 1rem 0;">
   <label style="display: flex; gap: 0.5rem; align-items: center;">
